backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
mysqlclient==2.2.7
requests==2.31.0
python-dotenv==1.0.1
django-autoslug==1.9.9
python-jose==3.3.0
drf-yasg==1.21.6
setuptools==80.9.0